Attempt to upgrade Jackson
This is my partially successful upgrade of the Jackson libraries. All the tests pass but one, but I have no idea why that one test doesn't pass.

I really need to move on to other things, so I'm going to save this work in a branch and come back to it later if I have time.

There's also a lot of clutter and cruft and commented out code in this because I'm obviously flailing a bit here. Sighz.