Add Cypress plugin for linter
https://typescript-eslint.io/rules/no-require-imports/
There was a complaint from ESLint about the use of require in the `cypress.config.ts` file.

There is a Cypress plugin for ESLint https://github.com/cypress-io/eslint-plugin-cypress and this approach to adding it to the `.eslintrc.json` is suggested in the history of the project. The project claims you need to be using ESLint 9 or greater, which we probably should look into, and it might mean that this plugin would need to be updated manually in that migration process. I have a PR for that migration as well:
https://github.com/UMM-CSci-3601/3601-iteration-template/pull/1664