Fix use of CMD to follow DeepSource suggestion
DeepSource tells me:
When using the plain text version of passing arguments, signals from the OS are not correctly passed to the executables, which is in the majority of the cases what you would expect.

These points shall always be taken care of:

CMD should almost always be used in the form of CMD [“executable”, “param1”, “param2”…]
The shell form prevents any CMD or run command line arguments from being used, but has the disadvantage that your ENTRYPOINT will be started as a subcommand of /bin/sh -c, which does not pass signals. This means that the executable will not be the container’s PID 1 - and will not receive Unix signals - so your executable will not receive a SIGTERM from docker stop .

Read more about these best practices here: https://docs.docker.com/build/building/best-practices/